Added support for quests in nebula sectors and the last stand
<nebulaQuest force="false" clear="false" nebulaEvent=""/>
<lastStandQuest noBoss="false"/>

These tags should be placed in the actual quest event (not the quest giver event, but the event that happens at the beacon marked "quest")

nebulaQuest allows a quest to spawn in nebula beacons. force="true" prevents the quest from spawning outside nebula beacons. clear="true" prevents the nebula background from being replaced if the quest spawns in a nebula beacon. nebulaEvent specifies the name of a replacement event if a nebula beacon is selected.

lastStandQuest allows the quest to spawn in sector 8 (or later), including the last stand. noBoss="true" prevents the quest from spawning in the actual last stand, but allows for quests in s8+ secret sectors.