Added current sector only quests
<noNextSectorQuest aggressive="false"/>

If the quest fails to spawn in the current sector, it will not be spawned at all. Setting aggressive="true" will allow a more aggressive attempt to spawn the quest if the initial attempt fails, by ignoring the player's and fleet's location.

The text entry in <text name="no_time_sector"> will be displayed if the quest fails to spawn.