Boss Fleet Protection for Events
Allows specifying events that are immune to being taken over by the fleet in sector 8 (either until no other beacons are left, or indefinitely). This can be used to prevent the merchant from being overtaken in sector 8 before he can be reached. This does not protect beacons overtaken by the Flagship.

Events can also be specified to be able to move to a new location if the fleet would overtake it (either preferring a nearby beacon or picking any beacon). This can be used to allow the travelling merchant to move to a new location if he is in the path of the boss (or if the boss would spawn on the same beacon as him).