systemPowerScaling now checks boost owner
systemPowerScaling was previously checking the ship of the crewmember whose stat was being modified, rather than checking the ship owning the stat boost. Now it checks the ship carrying the augment for augment boosts, and the ship owning the crewmember providing the boost for crew boosts.

Also modified Global::GetShipManager to modify how the enemy ship manager is retrieved in order to fix issues with neutral ships. Hopefully no new crashes.