Add ActivatedPower struct
The goal of this work is to restructure crew abilities into their own class. Currently there is a class for the ability definition (what defines the attributes of the crew ability) but the state of the crew ability itself is stored in both CrewMember_Extend and CrewAnimation_Extend.

Making the ability state its own class will allow for better encapsulation of crew abilities as well as pave the way for new features, such as allowing a single crewmember to possess an arbitrary number of abilities, instead of only one ability.