Fix: German Translation
Fixed a UI Bug in the german translation. For some reason a text in the hangar was shifted to the left. I came up with the janky and primitive solution to embed it with its UI onto the hangar image. It is janky but works great and can not be seen by the player