Replacement implementation for Get_Drone_Subclass
Automatically downcasts SpaceDrone* instances to their final type when passing them to lua.

This is similar to what Get_Drone_Subclass did, except this approach applies more generally to SpaceDrone* as a type and not just the result of a vector<SpaceDrone*> indexing operation.

NOTE: This does not apply to objects passed via SWIG_NewPointerObj (such as within InternalEvents::CONSTRUCT_SPACEDRONE, so Get_Space_Drone usage for these cases will no longer work.
However, downcasting would be a bug anyway, so this is unlikely to b e an issue.