Fixed crash upon starting a game
Previously the game would crash upon starting the game because of a BAD_ACCES  happening in the unmodified vanilla code from SystemControl::IsSystemTapped meaning that there has to be some sort of data corruption caused by HS code earlier. I figured out that skipping our reimplementation of SystemControl::CreateSystemBoxes fixed the crash, after days of debugging I finally managed to find the culprit which was the WeaponSystemBox aswell as the clonebox constructor hook adresses, which were actually the one for their destructors which explains why an illegal memory read happened.

- WeaponSystemBox.zhl fiexd WeaponSystemBox::constructor signature
- CloneBox.zhl. fixed CloneBox::constructor signature
- ShipSystem.zhl shortened a few signatures
- MindBox.zhl shortened the constructor signature